Fix compilation errors for APK build
- Add missing voice_service.dart with flutter_tts
- Fix FFI bindings to use proper Utf8 types from ffi package
- Fix storage_service.dart - remove ApiConfig references
- Fix chat_screen.dart - remove status parameter from TypingIndicator
- Fix home_screen.dart - remove isHighlighted parameter
- Fix model_manager_screen.dart - cast double to int for sizeBytes
- Fix local_llm_service.dart - add explicit type parameter for runInference
- Update android NDK version to 28.2.13676358
- Simplify llama_bridge.cpp to match FFI signatures

+++ lib/screens/model_manager_screen.dart
@@ -105,7 +105,7 @@ class _ModelManagerScreenState extends State<ModelManagerScreen> {
         id: 'phi-2-q4',
         name: 'Phi-2',
         description: 'Microsoft\'s small but powerful model',
-        sizeBytes: 1.6 * 1024 * 1024 * 1024,
+        sizeBytes: (1.6 * 1024 * 1024 * 1024).toInt(),
         parameters: 2700,
         quantization: 'Q4_K_M',
         downloadUrl: 'https://huggingface.co/TheBloke/phi-2-GGUF/resolve/main/phi-2.Q4_K_M.gguf',
